Adding "unknown total" behaviour
Pass in `null` as the total count to render next/previous links
without having to perform a `count()` on all items in the database.
Set `currentFetched` to the number of items you have fetched for
this request so that the paginator can guess if there is a next
page available.

E.g. Fetch maximally 21 items from the database while you only
want to display 20 items per page. Pass in the count of actually
fetched items as `currentFetched`, strip away the last one from
the items to be rendered, then echo the pagination links.